תכנות/ORCHESTRATOR/orchestrator/requirements.txt
google-generativeai
fastapi
uvicorn
sqlmodel
aiosqlite
typer
python-dotenv
pyyaml
rich
httpx
